refactor: move the Artist, Genres and Home graphs to the navgraph package
