+++ app/src/main/java/uk/udemy/recordshop/ui/navigation/NavigationGraph.kt
@@ -8,6 +8,9 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
 import androidx.navigation.NavHostController
 import kotlinx.coroutines.CoroutineScope
+import uk.udemy.recordshop.ui.navigation.navgraphs.artistGraph
+import uk.udemy.recordshop.ui.navigation.navgraphs.genresGraph
+import uk.udemy.recordshop.ui.navigation.navgraphs.homeGraph
 
 // The Root NavGraph
 @Composable
