+++ app/src/main/java/uk/udemy/recordshop/ui/navigation/navgraphs/GenresGraph.kt
@@ -1,10 +1,12 @@
-package uk.udemy.recordshop.ui.navigation
+package uk.udemy.recordshop.ui.navigation.navgraphs
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
 import uk.udemy.recordshop.ui.genres.GenresScreen
+import uk.udemy.recordshop.ui.navigation.Screens
+import uk.udemy.recordshop.ui.navigation.Tabs
 
 fun NavGraphBuilder.genresGraph(
     navController: NavController
