+++ app/src/main/java/uk/udemy/recordshop/ui/navigation/navgraphs/HomeGraph.kt
@@ -1,4 +1,4 @@
-package uk.udemy.recordshop.ui.navigation
+package uk.udemy.recordshop.ui.navigation.navgraphs
 
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.LaunchedEffect
@@ -14,6 +14,8 @@ import uk.udemy.recordshop.ui.addOrEditAlbum.AddOrEditAlbumScreen
 import uk.udemy.recordshop.ui.addOrEditAlbum.AddOrEditAlbumViewModel
 import uk.udemy.recordshop.ui.home.HomeScreen
 import uk.udemy.recordshop.ui.home.HomeViewModel
+import uk.udemy.recordshop.ui.navigation.Screens
+import uk.udemy.recordshop.ui.navigation.Tabs
 import uk.udemy.recordshop.ui.viewAlbum.ViewAlbumScreen
 import uk.udemy.recordshop.ui.viewAlbum.ViewAlbumViewModel
